+++ unit_test/tstFieldManager.hpp
@@ -32,8 +32,8 @@ void uniformTest()
 {
     // Get inputs for mesh.
     InputParser parser( "uniform_mesh_test_1.json", "json" );
-    Kokkos::Array<float,6> global_box = { -10.0, -10.0, -10.0,
-                                          10.0, 10.0, 10.0 };
+    Kokkos::Array<double,6> global_box = { -10.0, -10.0, -10.0,
+                                           10.0, 10.0, 10.0 };
     int minimum_halo_size = 0;
 
     // Make mesh.
@@ -139,8 +139,8 @@ void adaptiveTest()
 {
     // Get inputs for mesh.
     InputParser parser( "adaptive_mesh_test_1.json", "json" );
-    Kokkos::Array<float,6> global_box = { -10.0, -10.0, -10.0,
-                                          10.0, 10.0, 10.0 };
+    Kokkos::Array<double,6> global_box = { -10.0, -10.0, -10.0,
+                                           10.0, 10.0, 10.0 };
     int minimum_halo_size = 1;
     double cell_size = 0.5;
